refactor(GAME): Reorganize backend files for game managment
